prevent inline assembly instruction reordering in muldiv.h
It was theoretically possible that the compiler could reorder the "adc" (add with carry) instructions in the signed muladddiv functions to before the "add" instructions which set the requisite carry flag. This commit introduces a faux input operand on the "adc" instructions to force a data dependency on the output operand of the "add" instructions, thus forcing the compiler to emit these instructions in the proper order. This is a preventative measure; there is no evidence that the compiler was actually reordering these instructions in practice.

This commit also adds "cc" (condition code) as a clobber operand to preclude the compiler from assuming that the CPU condition flags are preserved across calls to muldiv/muladddiv.