add const and pure attributes to various functions
Though in theory this could be a performance optimization, the affected functions are simple enough that link-time optimization would inline and hoist their bodies out of loops anyway, so in practice these attributes won't improve code generation. However, the presence of these attributes does allow the compiler to emit a warning if the return value from any of these functions is unused, and that's a helpful diagnostic.