do not delete Selectable instance in Selectable::pump() exception handler
It cannot be guaranteed (and in fact it is unlikely) that the Selector holds the only pointer to the Selectable, so it does not make sense for the Selector to assume ownership of the Selectable. To avoid the need for Selectable::pump() to handle exceptions escaping from Selectable::selected(…), declare Selectable::selected() as noexcept.