io: introduce Readable, Writable, Flushable mix-in classes using CRTP
One consequence of this change is that FileDescriptor no longer implements Source and Sink, reducing sizeof(FileDescriptor) from 24 bytes down to 4 bytes. To enable a FileDescriptor to be passed to functions expecting these interfaces, adapter classes FileSource and FileSink are introduced. The expectation is that instances of these adapter classes may often be shorter-lived than the FileDescriptor they reference.