fd: relay void return value from forwarding functions (code style)
When a function's main purpose and final operation is to call another function, and that other function's return type is void, return the void value of the function call expression. C++ allows this, and it's a great way to ensure that any future addition of a non-void return type to the called function does not go unnoticed by the forwarding function.